tcp output moved to a package
